#384: moved SpringBootApp and SpringBootBatchApp to server
