+++ samples/core/src/test/java/io/oasp/gastronomy/restaurant/SpringBootBatchTestApp.java
@@ -3,24 +3,29 @@ package io.oasp.gastronomy.restaurant;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.actuate.autoconfigure.EndpointAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.security.SecurityAutoConfiguration;
+import org.springframework.boot.autoconfigure.security.SecurityFilterAutoConfiguration;
 import org.springframework.boot.orm.jpa.EntityScan;
 import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
 
 import io.oasp.module.jpa.dataaccess.api.AdvancedRevisionEntity;
 
-//@SpringBootApplication(exclude = { SecurityAutoConfiguration.class, SecurityFilterAutoConfiguration.class })
-@SpringBootApplication(exclude = { EndpointAutoConfiguration.class })
+@SpringBootApplication(exclude = {
+		EndpointAutoConfiguration.class,
+		SecurityAutoConfiguration.class,
+		SecurityFilterAutoConfiguration.class,
+		})
 @EntityScan(basePackages = { "io.oasp.gastronomy.restaurant" }, basePackageClasses = { AdvancedRevisionEntity.class })
-@EnableGlobalMethodSecurity(jsr250Enabled = true)
-public class SpringBootApp {
+@EnableGlobalMethodSecurity(jsr250Enabled = false)
+public class SpringBootBatchTestApp {
 
   /**
    * Entry point for spring-boot based app
    *
    * @param args - arguments
    */
-  public static void main(String[] args) {
+	  public static void main(String[] args) {
 
-    SpringApplication.run(SpringBootApp.class, args);
-  }
+		    SpringApplication.run(SpringBootTestApp.class, args);
+	  }
 }
