+++ samples/server/src/main/java/io/oasp/gastronomy/restaurant/general/configuration/ServletInitializer.java
@@ -1,11 +1,11 @@
 package io.oasp.gastronomy.restaurant.general.configuration;
 
-import io.oasp.gastronomy.restaurant.SpringBootApp;
-
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.context.web.SpringBootServletInitializer;
 import org.springframework.context.annotation.Configuration;
+
+import io.oasp.gastronomy.restaurant.SpringBootApp;
 
 /**
  * This auto configuration will be used by spring boot to enable traditional deployment to a servlet container. You may
